requirements.txt
# ── Server ────────────────────────────────────────────────────────────────────
fastapi
uvicorn
anyio
python-multipart

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy
psycopg2-binary

# ── Auth & Security ───────────────────────────────────────────────────────────
python-jose[cryptography]
passlib[argon2]
argon2-cffi

# ── Settings ──────────────────────────────────────────────────────────────────
pydantic[email]>=2.7.4,<3.0.0
pydantic-settings
python-dotenv

# ── Redis ─────────────────────────────────────────────────────────────────────
redis>=5.0.0

# ── HTTP client (mail service) ────────────────────────────────────────────────
requests

# ── LangChain ─────────────────────────────────────────────────────────────────
langchain>=0.2.11
langchain-community>=0.2.11
langchain-core>=0.2.11
langchain-pinecone>=0.1.3
langchain-huggingface>=0.0.3

# ── Vector DB ─────────────────────────────────────────────────────────────────
pinecone-client>=3.0.0

# ── LLM ───────────────────────────────────────────────────────────────────────
groq==0.13.0

# ── Embeddings ────────────────────────────────────────────────────────────────
huggingface-hub>=0.19.0
sentence-transformers>=2.2.0

# ── Document parsing ──────────────────────────────────────────────────────────
PyPDF2==3.0.1
python-docx==1.1.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest
